Refactor : game 페이지 로직, 스타일, jsx 분리하는 리팩토링 적용
상세:
- jsx 부분을 presenter.jsx로 분리하여 넣고 export
- presenter는 필요한 부분을 Game/index.js 로부터 props의 형태로 전달 받아 렌더
- materialui의 make 스타일 부분도 index에서 생성하여 사용하고 presenter로 전달

328
